backups/2023-12-21_1642/requirements.txt
flask
python-dotenv
authlib
qrcode
reportlab
beautifulsoup4
pandas
pytz
requests
